backend/requirements.txt
django==5.2.0
djangorestframework==3.15.0
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.0.0
numpy==1.26.4
pandas==2.2.1
websockets==12.0